Fix function signatures and build errors across subsystems
Time subsystem:
- TIME_$SET_CPU_LIMIT: Changed to (clock_t*, int8_t*, status_$t*)
- TIME_$SET_ITIMER: Changed to 6 params with correct types
- TIME_$SET_VECTOR: Changed to void params
- TIME_$WAIT/WAIT2: Fixed param counts and types
- SUB48: Changed return type to int8_t (returns sign indicator)

TTY subsystem:
- Fixed TTY_$K_SET_INPUT_FLAG, TTY_$K_SET_OUTPUT_FLAG param types
- Fixed TTY_$K_SET_FUNC_CHAR, TTY_$K_INQ_FUNC_CHAR, TTY_$K_ENABLE_FUNC

Other fixes:
- Fixed PROC2_UID access pattern (use pointer arithmetic not array)
- Fixed uid/gen.c clock type usage
- Fixed proc1 assembly branch offsets (bra.s -> bra.w)
- Removed duplicate Illegal_PID_Err definition
- Added -nostdlib -ffreestanding to linker flags
- Created misc/crash_system.h and misc/string.h headers

+++ domain_os/proc2/get_cr_rec.c
@@ -40,7 +40,7 @@ void PROC2_$GET_CR_REC(uint32_t *ec_handle, uid_t *parent_uid, uid_t *proc_uid,
     /* Calculate process index from EC1 address:
      * index = ((ec1_addr - EC1_ARRAY_BASE) / EC1_ENTRY_SIZE) + 1
      */
-    proc_idx = (int16_t)(((uintptr_t)ec1_addr - EC1_ARRAY_BASE) / EC1_ENTRY_SIZE) + 1;
+    proc_idx = (int16_t)(((uintptr_t)ec1_addr - (uintptr_t)EC1_ARRAY_BASE) / EC1_ENTRY_SIZE) + 1;
 
     if (status != status_$ok) {
         *status_ret = status_$proc2_uid_not_found;
